Home: move the actions data to the demo state
The home cards are now opening the corresponding apps. For now, they
only request to open an instance. Later, a context will be added to
provide a way to navigate inside of an app instance.

The MenuPanel has been changed to reflect this: instead of changing the
path directly, it also requests to open an instance. The `openApp`
method of `App` is responsible of updating the path.