Add the Home app (#74)
This commit adds the Home app.

It also changes the way path updates are handled: instead of changing the
path directly, it also requests to open an instance. The `openApp(appId, instanceId)`
method of `App` is responsible of updating the path.

For now, the Home cards only request to open an instance. Later, a context will be added to
provide a way to navigate inside of an app instance.