Burned permissions (#415)
With `aragonOS@4` we have have the standard notion of a "burned" permission manager, i.e. a permission with its manager thrown away so that it's forever frozen (from the state it was in when it was "burned").

This fixes the app crashing when no `allowedEntities` exist, but the manager does (likely because the permission's been completely thrown away), and also adds detection for these "burned" managers. I've opted to use `Discarded` as the label name.