Fix an onboarding crash happening with specific web3 provider setups
When calling getBalance(), it was possible to sometimes get another
value than an big integer as a string.

Having `null` as a result, and passing it to the BN.js constructor,
could lead to an infinit loop [1].

To prevent this issue to happen again:

- In the app, `balance` is now always represented by a BN.js instance.
To represent an unknown balance, `new BN(-1)` is now used rather than `null`.

- The result of getbalance() is now filtered to ensure that we are
passing an integer to BN.js. Otherwise, we pass "-1".

[1] https://github.com/indutny/bn.js/issues/186