Upgrade to react-spring 6
The module src/springs.js is used instead of coming from Aragon UI. This
is because the springs coming with the current version of Aragon UI
(0.27) are not compatible with react-spring 6. The next Aragon UI
version will fix that, and the module src/spring.js will be removed at
this point.

This commit also fixes a number of other things:

- Home app: use <AppView>, better centering, slightly smaller cards,
  better transition.
- Menu apps loading: fix the multiple instances issue, make the
  animation simpler and faster to give a similar feel than what we have on
  the permissions app or the navigation bar in Aragon UI.
- Other minor tweaks related to the way we manage transitions.