ZoomableCards: first version
ZoomableCards (temporary name) is a component that allows to easily
create layouts similar to the Survey app, with cards that can be
expanded into a full screen layout composed of the expanded card and an
optional sidebar, entering in the screen from the right side (and whose
transition starts slightly after).

At smaller viewport sizes, the content of the sidebar moves inside the
expanded card itself.