feat: restart app worker when upgraded (#688)
Coupled with https://github.com/aragon/aragon.js/pull/267, this allows the client to restart an upgraded app's worker in-place.

We reset the cache on the upgrade as well, as there may be incompatibilities in the reduced state between different worker versions (as well as paired with their frontends). Note that app frontends are handled mostly transparently as we re-load their URL into the iframe upon navigation.