MenuPanel: adjust for new styles (#923)
Tweaks the menu panel and organization switcher to the refreshed designs. Also refactor and rewire CombinedPanel and autocollapse mode.

Also simplifies the aragon.js error handling a bit, to push everything to the global error handler for now. We used to have reload states for the apps in case they weren't able to load (e.g. because of a bad IPFS connection) but these errors are not propagated the same way from aragon.js anymore and we should later revisit how to properly handle these cases (#84, #361).