Configuration: move prefix from REACT_APP to ARAGON (#1122)
The REACT_APP prefix was coming from Create React App that we don’t use anymore. Using a prefix like ARAGON is probably better in the long term (and less verbose!). Compatibility with REACT_APP prefixes could be maintained for some time as it only impact a few areas.

Also removes ARAGON_DEMO_DAO which is not used anymore.