AccountModule: copy address feature
More as a proof-of-concept, as this solution hacks IdentityBadge, while
it should be a variant of IdentityBadge. The IdentityBadge and copy icon
could then be inside a single clickable element.

Also, a useCopyToClipboard helper is added - if effective, it should be
moved to a more universal place for reuse, e.g. in ShareModal component.