Account module improvements (#1137)
AccountModule: show popup with provider & net info.

- refactor; provider & net info
- handle network mismatch error
- create getWeb3Instances helper
- TopBar: styling fix, so that Account Module has more surface

AccountModule: copy address feature.

More as a proof-of-concept, as this solution hacks IdentityBadge, while
it should be a variant of IdentityBadge. The IdentityBadge and copy icon
could then be inside a single clickable element.

Also, a useCopyToClipboard helper is added - if effective, it should be
moved to a more universal place for reuse, e.g. in ShareModal component.

Providers: handle unknown provider.

This case was not handled, so the app would crash if provider was 'unknown'

Networks: add short names + getNetworkByChainId().

Connected account: move from networkId to chainId.

Add a delay before displaying the button / account module.